+++ src/datagen.rs
@@ -17,7 +17,8 @@ use crate::{
         Board, GameOutcome,
     },
     definitions::{depth::Depth, MEGABYTE},
-    searchinfo::{SearchInfo, SearchLimit},
+    searchinfo::SearchInfo,
+    timemgmt::SearchLimit,
     tablebases::{self, probe::WDL},
     threadlocal::ThreadData,
     transpositiontable::TT,
